+++ globalVariables.h
@@ -14,16 +14,16 @@
 #include <windows.h> // for MS Windows
 #include <GL/glut.h> // GLUT, include glu.h and gl.h
 
-float nearPlane = 0.0f;
-float farPlane = 0.0f;
-float fovY = 0.0f;
-float aspectRatio = 0.0f;
+double nearPlane = 0.0f;
+double farPlane = 0.0f;
+double fovY = 0.0f;
+double aspectRatio = 0.0f;
 int recursionLevel = 0;
 int numPixels = 0;
 int checkerboardWidth = 0;
-float checkerboardAmbientCoeff = 0.0f;
-float checkerboardDiffuseCoeff = 0.0f;
-float checkerboardReflectionCoeff = 0.0f;
+double checkerboardAmbientCoeff = 0.0f;
+double checkerboardDiffuseCoeff = 0.0f;
+double checkerboardReflectionCoeff = 0.0f;
 
 char title[] = "3D Shapes";
 
@@ -42,8 +42,8 @@ void resetCamera()
 int prevMouseX, prevMouseY;    // Previous mouse coordinates
 bool mouseInitialized = false; // Flag to check if the mouse has been initialized
 
-GLfloat ScreenWidth = 640.0f;
-GLfloat ScreenHeight = 480.0f;
+GLdouble ScreenWidth = 640.0f;
+GLdouble ScreenHeight = 480.0f;
 
 /* All objects for scene */
 std::vector<Sphere> spheres;
